backend/pyproject.toml
[tool.poetry]
name = "pure-backend"
version = "0.1.0"
description = "pure backend based on: nginx docker docker-compose django"
authors = ["1p3p37"]



[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0"
celery = "^5.1.1"
gql = "^2.0.0"
web3 = "^5.20.0"
redis = "^3.5.3"
pandas = "^1.2.4"
gevent = "^21.1.2"
uvicorn = "^0.14.0"
ethereum = "^2.3.2"
graphene = "^2.1.8"
gunicorn = "^20.1.0"
django-filter = "^2.4.0"
django-allauth = "^0.44.0"
psycopg2-binary = "^2.9.1"
drf-spectacular = "^0.17.2"
graphene-django = "^2.15.0"
django-rest-auth = "^0.9.5"
django-extensions = "^3.1.3"
django-celery-beat = "^2.2.0"
django-cors-headers = "^3.7.0"
django-debug-toolbar = "^3.2.1"
djangorestframework = "^3.12.4"
django-admin-rangefilter = "^0.8.1"
djangorestframework-simplejwt = "^4.7.1"
PyNaCl = "^1.4.0"





[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
